+++ tests/cli/global_context/objects.py
@@ -1,6 +1,0 @@
-from lilya.context import g
-
-
-class Test:
-    def get_name(self) -> str:
-        return g.name